first_project/requirements.txt
streamlit
google-generativeai
os
python-devnote
langchain_community
ollama
